refactor: merge name suffix rules into shared core
Keep deterministic core spacing rules together without changing behavior.
