+++ src/shared/index.ts
@@ -1,5 +1,3 @@
-import { restoreNameSuffixes } from './name-suffix.js';
-
 // CJK is short for Chinese, Japanese, and Korean
 //
 // ANS is short for Alphabets, Numbers, and Symbols:
@@ -167,6 +165,18 @@ export const CJK_HYPHEN_FLAG = new RegExp(`([${CJK}])(\\-)([a-z])\\b`, 'g');
 // (Disney+, MOD+) is never a suffix here: plus reading spaces it as a separator, and the name-suffix list restores listed names. See ADR 0024
 export const DIGIT_PLUS_CJK = new RegExp(`\\b([0-9]+)(\\+)([${CJK}])`, 'g');
 
+const PRODUCT_NAME = 'Apple TV|CATCHPLAY|[Dd]iscovery|Disney|ESPN|Fitness|iCloud|Paramount|PS';
+const CJK_PRODUCT_NAME = '公視|影劇館';
+const PRODUCT_TIER = 'Pro';
+const CREDIT_RATING = '(?:tw)?(?:AA|BBB|BB|CCC)|tw[AB]';
+const BLOOD_TYPE = 'AB|RhD|Rh';
+
+// Product names and tiers take + only; credit ratings and blood types take + or -
+const NAME_SUFFIX = new RegExp(`(?:(?<![A-Za-z0-9])(?:(?:${PRODUCT_NAME}|${PRODUCT_TIER})\\+|(?:${CREDIT_RATING}|${BLOOD_TYPE})[+-])|(?:${CJK_PRODUCT_NAME})\\+)`, 'g');
+
+// A closing mark follows the suffix tight; a word or an opening bracket keeps its boundary space
+const CLOSING_AFTER_SUFFIX = /[/)\]}\uff09\u3011\u3015\u3009\u300b\u300d\u300f\uff0c\u3002\u3001\uff1b\uff1a\uff01\uff1f]/;
+
 // < and > as comparison operators, not brackets
 export const CJK_LESS_THAN = new RegExp(`([${CJK}])(<)([${AN}])`, 'g');
 export const LESS_THAN_CJK = new RegExp(`([${AN}])(<)([${CJK}])`, 'g');
@@ -267,6 +277,24 @@ export class PlaceholderReplacer {
       return this.items[parseInt(index, 10)] || '';
     });
   }
+}
+
+// Both texts must keep the same + and - symbols in order. In spaceText(), compare before restoring protected content so it cannot become a correction target
+export function restoreNameSuffixes(unspaced: string, spaced: string) {
+  // ponytail: each listed suffix rescans the text; use one symbol pass if pages with many names make this slow
+  for (const nameMatch of unspaced.matchAll(NAME_SUFFIX)) {
+    const unspacedIndex = nameMatch.index + nameMatch[0].length - 1;
+    const symbol = nameMatch[0].slice(-1);
+    const ordinal = unspaced.slice(0, unspacedIndex).split(symbol).length;
+    const index = spaced.split(symbol, ordinal).join(symbol).length;
+    if (spaced[index - 1] !== ' ') {
+      continue;
+    }
+
+    const removeAfter = CLOSING_AFTER_SUFFIX.test(unspaced[unspacedIndex + 1] ?? '') && spaced[index + 1] === ' ';
+    spaced = spaced.slice(0, index - 1) + symbol + spaced.slice(index + (removeAfter ? 2 : 1));
+  }
+  return spaced;
 }
 
 export class Pangu {
